+++ api/src/Elasticsuite/Standard/src/Search/Elasticsearch/Request/Query/Terms.php
@@ -25,7 +25,7 @@ class Terms implements QueryInterface
 {
     private ?string $name;
 
-    private int $boost;
+    private float $boost;
 
     private array $values;
 
@@ -37,13 +37,13 @@ class Terms implements QueryInterface
      * @param string|bool|array $values Search values. String are exploded using the comma as separator
      * @param string            $field  Search field
      * @param ?string           $name   Name of the query
-     * @param int               $boost  Query boost
+     * @param float             $boost  Query boost
      */
     public function __construct(
         string|bool|array $values,
         string $field,
         ?string $name = null,
-        int $boost = QueryInterface::DEFAULT_BOOST_VALUE
+        float $boost = QueryInterface::DEFAULT_BOOST_VALUE
     ) {
         if (!\is_array($values) && \is_string($values)) {
             $values = explode(',', $values);
@@ -76,7 +76,7 @@ class Terms implements QueryInterface
     /**
      * {@inheritDoc}
      */
-    public function getBoost(): int
+    public function getBoost(): float
     {
         return $this->boost;
     }
